+++ DuckDuckGo/Assets.xcassets/AddFavorite.imageset/addFavorite_dark.pdf
