--- DuckDuckGo/Assets.xcassets/AddFavoriteLight.imageset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "images" : [
-    {
-      "idiom" : "universal",
-      "filename" : "addFavorite_light.pdf"
-    }
-  ],
-  "info" : {
-    "version" : 1,
-    "author" : "xcode"
-  }
-}
