+++ DuckDuckGo/Assets.xcassets/LogoDarkText.imageset/Contents.json
@@ -2,7 +2,7 @@
   "images" : [
     {
       "idiom" : "universal",
-      "filename" : "updatedLogoGrayText.pdf"
+      "filename" : "light_portrait_logo.pdf"
     },
     {
       "idiom" : "universal",
