+++ DuckDuckGo/Assets.xcassets/LogoLightText.imageset/Contents.json
@@ -2,7 +2,7 @@
   "images" : [
     {
       "idiom" : "universal",
-      "filename" : "updatedLogoWhiteText.pdf"
+      "filename" : "dark_portrait_logo.pdf"
     },
     {
       "idiom" : "universal",
