+++ DuckDuckGo/HomeViewController.swift
@@ -57,6 +57,7 @@ class HomeViewController: UIViewController {
         
         applyTheme(ThemeManager.shared.currentTheme)
         
+
     }
     
     func refresh() {
@@ -230,193 +231,4 @@ extension HomeViewController: Themable {
         collectionView.reloadData()
         view.backgroundColor = theme.backgroundColor
     }
-}
-
-class NavigationSearchHomeCell: ThemableCollectionViewCell {
-    
-    @IBOutlet weak var imageView: UIImageView!
-    
-    override func decorate(with theme: Theme) {
-        switch theme.currentImageSet {
-        case .light:
-            imageView.image = UIImage(named: "LogoDarkText")
-        case .dark:
-            imageView.image = UIImage(named: "LogoLightText")
-        }
-    }
-    
-}
-
-class CenteredSearchHomeCell: ThemableCollectionViewCell {
-    
-    @IBOutlet weak var imageView: UIImageView!
-    @IBOutlet weak var searchBackground: RoundedRectangleView!
-    @IBOutlet weak var promptText: UILabel!
-    @IBOutlet weak var searchLoupe: UIImageView!
-
-    private lazy var tapGesture = UITapGestureRecognizer(target: self, action: #selector(onTap))
-    
-    var tapped: ((CenteredSearchHomeCell) -> Void)?
-    
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        searchBackground.addGestureRecognizer(tapGesture)
-    }
-    
-    override func decorate(with theme: Theme) {
-        searchBackground.borderColor = theme.searchBarBackgroundColor
-        searchBackground.backgroundColor = theme.searchBarBackgroundColor
-        searchLoupe.tintColor = theme.barTintColor
-        promptText.textColor = UIColor.greyish // TODO should this be a themeable color (if so also apply to omnibar)
-                
-        switch theme.currentImageSet {
-        case .light:
-            imageView.image = UIImage(named: "LogoDarkText")
-        case .dark:
-            imageView.image = UIImage(named: "LogoLightText")
-        }
-    }
-
-    @objc func onTap() {
-        print("***", #function)
-        tapped?(self)
-    }
-    
-}
-
-class AddFavoriteHomeCell: ThemableCollectionViewCell {
-    
-    @IBOutlet weak var plusImage: UIImageView!
-    @IBOutlet weak var plusBackground: UIView!
-    
-    override func decorate(with theme: Theme) {
-        
-        switch theme.currentImageSet {
-        case .light:
-            plusImage.image = UIImage(named: "AddFavoriteLight")
-            plusBackground.backgroundColor = UIColor.mercury
-            
-        case .dark:
-            plusImage.image = UIImage(named: "AddFavoriteDark")
-            plusBackground.backgroundColor = UIColor.black
-        }
-    }
-    
-}
-
-class FavoriteHomeCell: ThemableCollectionViewCell {
-    
-    @IBOutlet weak var titleLabel: UILabel!
-    @IBOutlet weak var iconLabel: UILabel!
-    @IBOutlet weak var iconBackground: UIView!
-    @IBOutlet weak var iconImage: UIImageView!
-
-    var onDelete: (() -> Void)?
-    var onEdit: (() -> Void)?
-    
-    private var link: Link?
-    private var theme: Theme?
-    
-    struct Actions {
-        static let delete = #selector(FavoriteHomeCell.doDelete(sender:))
-        static let edit = #selector(FavoriteHomeCell.doEdit(sender:))
-    }
-    
-    override var canBecomeFirstResponder: Bool {
-        return true
-    }
-    
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        iconBackground.layer.shadowRadius = 1
-        iconBackground.layer.shadowOffset = CGSize(width: 0, height: 1)
-        iconBackground.layer.shadowColor = UIColor.black.cgColor
-        iconBackground.layer.shadowOpacity = 0.12
-    }
-    
-    @objc func doDelete(sender: Any?) {
-        print("***", #function)
-        onDelete?()
-    }
-    
-    @objc func doEdit(sender: Any?) {
-        print("***", #function)
-        onEdit?()
-    }
-    
-    override func canPerformAction(_ action: Selector, withSender sender: Any?) -> Bool {
-        print("***", #function, action)
-        return [ Actions.delete, Actions.edit ].contains(action)
-    }
-    
-    override func decorate(with theme: Theme) {
-        super.decorate(with: theme)
-        self.theme = theme
-        
-        if let link = link {
-            updateFor(link: link)
-        }
-    }
-    
-    func updateFor(link: Link) {
-        self.link = link
-        
-        let host = link.url.host?.dropPrefix(prefix: "www.") ?? ""
-        iconLabel.text = "\(host.capitalized.first ?? " ")"
-        
-        titleLabel.text = link.title
-        
-        iconImage.isHidden = true
-        iconLabel.isHidden = false
-        
-        iconBackground.backgroundColor = host.color
-        
-        if let domain = link.url.host {
-            let resource = AppUrls().faviconUrl(forDomain: domain)
-            iconImage.kf.setImage(with: resource, placeholder: nil, options: nil, progressBlock: nil) { [weak self] image, error, _, _ in
-                guard error == nil else { return }
-                guard let image = image else { return }
-                guard image.size.width > 16 else { return }
-                self?.applyFavicon(image)
-            }
-        }
-        
-    }
-      
-    func applyFavicon(_ image: UIImage) {
-        iconLabel.isHidden = true
-        
-        iconImage.isHidden = false
-        iconImage.contentMode = image.size.width >= 64 ? .scaleAspectFit : .center
-        iconImage.layer.masksToBounds = true
-        iconImage.layer.cornerRadius = 8
-        
-        guard let theme = theme else { return }
-        switch theme.currentImageSet {
-        case .dark:
-            iconBackground.backgroundColor = UIColor.charcoalGrey
-            
-        case .light:
-            iconBackground.backgroundColor = UIColor.white
-        }
-    }
-    
-}
-
-fileprivate extension String {
-    
-    var consistentHash: Int {
-        return self.utf8
-            .map { return $0 }
-            .reduce(5381) { ($0 << 5) &+ $0 &+ Int($1) }
-    }
-    
-    var color: UIColor {
-        let hash = consistentHash
-        let red = CGFloat((hash >> 0) & 0xFF)
-        let green = CGFloat((hash >> 8) & 0xFF)
-        let blue = CGFloat((hash >> 16) & 0xFF)
-        return UIColor(red: red / 255, green: green / 255, blue: blue / 255, alpha: 1.0)
-    }
-    
 }
