+++ DuckDuckGo/NavigationSearchHomeCell.swift
@@ -1,0 +1,24 @@
+//
+//  NavigationSearchHomeCell.swift
+//  DuckDuckGo
+//
+//  Created by Chris Brind on 28/11/2018.
+//  Copyright © 2018 DuckDuckGo. All rights reserved.
+//
+
+import UIKit
+
+class NavigationSearchHomeCell: ThemableCollectionViewCell {
+    
+    @IBOutlet weak var imageView: UIImageView!
+    
+    override func decorate(with theme: Theme) {
+        switch theme.currentImageSet {
+        case .light:
+            imageView.image = UIImage(named: "LogoDarkText")
+        case .dark:
+            imageView.image = UIImage(named: "LogoLightText")
+        }
+    }
+    
+}
